% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcpy.r
\docType{data}
\name{arcpy}
\alias{arcpy}
\title{ArcPy Module}
\format{An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.}
\usage{
arcpy
}
\description{
The arcpy module
}
\keyword{datasets}
