% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{arcpy_is_loaded}
\alias{arcpy_is_loaded}
\alias{set_arcpy_env}
\alias{unset_arcpy_env}
\title{ArcGIS Environment Loaded}
\usage{
arcpy_is_loaded()

set_arcpy_env(env_name)

unset_arcpy_env()
}
\arguments{
\item{env_name}{The name of the ArcGIS Pro Conda environment,
or the path to the ArcGIS Desktop Python binary.}
}
\description{
Check if ArcGIS Environment is Loaded
}
\keyword{internal}
