% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.r
\name{arcpy_version}
\alias{arcpy_version}
\title{Get Arcpy Version}
\usage{
arcpy_version(version, conda = "auto", channel = "esri", forge = TRUE)
}
\arguments{
\item{version}{The arcpy module version.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}

\item{channel}{An optional character vector of conda channels to include.
When specified, the \code{forge} argument is ignored. If you need to
specify multiple channels, including the conda forge, you can use
\verb{c("conda-forge", <other channels>)}.}

\item{forge}{Boolean; include the \href{https://conda-forge.org/}{conda-forge}
repository?}
}
\value{
A named list providing version numbers of arcpy and Python.
}
\description{
Verify the supplied arcpy module version and identify the required
Python version.
}
