% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{create_arcpy_env}
\alias{create_arcpy_env}
\title{Create ArcGIS Conda Environment}
\usage{
create_arcpy_env(
  env_name = "arcpy-env",
  arcgis_version = "",
  python_version = ""
)
}
\arguments{
\item{env_name}{The name of the Conda environment to create.
Default is \code{"arcpy_env"}.}

\item{arcgis_version}{The ArcGIS Pro version,
in format \verb{<major>.<minor>}.}

\item{python_version}{The Python version used by ArcGIS Pro,
in format \verb{<major>.<minor>}.}
}
\description{
Create a Conda environment with the "arcpy" module.
}
\details{
The Conda environment must be configured to match the
ArcGIS Pro version currently installed, and must be recreated when
ArcGIS Pro is updated.
}
