% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.r
\name{da_fields}
\alias{da_fields}
\title{List Attribute Table Fields}
\usage{
da_fields(table.path)
}
\arguments{
\item{table.path}{The file path to the table.}
}
\value{
A vector of field names.
}
\description{
Read attribute table field names with arcpy.da module.
}
\examples{
\dontrun{
arcpy$env$workspace = tempdir()
arcpy$env$scratchWorkspace = tempdir()
fc = arcpy$management$CopyFeatures(system.file("CA_Counties",
  "CA_Counties_TIGER2016.shp", package = "arcpy"), "CA_Counties")
da_fields(fc)
}

}
