% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.r
\name{da_update}
\alias{da_update}
\title{Update Table with arcpy.da}
\usage{
da_update(table.path, d)
}
\arguments{
\item{table.path}{The file path to the table.}

\item{d}{The data to write to \code{table.path}, with the same number 
of rows as the table. Column names must match field names 
of the table.}
}
\value{
(Invisible) The path to the table, i.e. \code{table.path}.
}
\description{
Update a table (e.g. attribute table of a layer) with the 
arcpy.da module.
}
\examples{
\dontrun{
layer = "path/to/table"
fields = c("VALUE", "NOTE")
d = da_read(layer, fields)
d["VALUE"] = d$VALUE + 5
d["NOTE"] = "modified"
da_update(layer, d)
}

}
