% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.r
\name{feature_from_arcpy}
\alias{feature_from_arcpy}
\title{Feature From ArcGIS Object}
\usage{
feature_from_arcpy(x, ...)
}
\arguments{
\item{x}{An ArcGIS vector object.}

\item{...}{Reserved for future expansion.}
}
\value{
A \code{sf} object.
}
\description{
Convert an ArcGIS object to an \code{sf} object.
}
\keyword{internal}
