% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.r
\name{feature_to_arcpy}
\alias{feature_to_arcpy}
\title{Feature To ArcGIS Object}
\usage{
feature_to_arcpy(x, ...)
}
\arguments{
\item{x}{An \code{sf} object.}

\item{...}{Reserved for future expansion.}
}
\value{
An ArcGIS vector object.
}
\description{
Convert an \code{sf} object to ArcGIS.
}
\keyword{internal}
