% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.r
\name{fields_exist}
\alias{fields_exist}
\title{Fields Exist}
\usage{
fields_exist(table.path, fields)
}
\arguments{
\item{table.path}{The file path to the table.}

\item{fields}{The field names.}
}
\value{
No return value; will produce an error if any of the
specified fields are not present in the table.
}
\description{
Check if specified fields are present in a table.
}
\keyword{internal}
