% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{find_ArcGIS}
\alias{find_ArcGIS}
\title{Find ArcGIS Installation}
\usage{
find_ArcGIS(pro = TRUE, required = TRUE)
}
\arguments{
\item{pro}{If \code{TRUE}, Search for ArcGIS Pro
Python distribution. Otherwise, search for ArcGIS
Desktop Python distribution.}
}
\value{
Path(s) to ArcGIS Python distributions.
}
\description{
Attempt to automatically locate an ArcGIS installation.
}
\seealso{
\code{\link{use_ArcGIS}}
}
