% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.r
\name{raster_from_arcpy}
\alias{raster_from_arcpy}
\title{Raster From ArcGIS Object}
\usage{
raster_from_arcpy(x, ...)
}
\arguments{
\item{x}{An ArcGIS raster object.}

\item{...}{Reserved for future expansion.}
}
\value{
A \code{SpatRaster} object.
}
\description{
Convert an ArcGIS raster object to a \code{SpatRaster} object.
}
\keyword{internal}
