% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.r
\name{raster_to_arcpy}
\alias{raster_to_arcpy}
\title{Raster To ArcGIS Object}
\usage{
raster_to_arcpy(x, ...)
}
\arguments{
\item{x}{A \code{SpatRaster} object.}

\item{...}{Reserved for future expansion.}
}
\value{
An ArcGIS Raster object.
}
\description{
Convert a \code{SpatRaster} object to ArcGIS.
}
\keyword{internal}
