% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcpy-methods.r
\name{RasterAlgebra}
\alias{RasterAlgebra}
\title{Arcpy Raster Algebra}
\arguments{
\item{a, b}{arcpy raster objects.}

\item{n}{non-negative integer.}
}
\description{
Operators for arcpy raster algebra.
}
\details{
The following operators are supported:

Arithmetic Operators:
\itemize{
\item \code{+}, \code{-}, and\code{/}
\item \code{**} and \code{^}
\item \code{\%\%} and \code{\%/\%}
}

Logical Operators:
\itemize{
\item \code{!}, \code{&}, \code{|}, and \code{xor}
\item \code{!=} and \code{==}
\item \code{<}, \code{<=}, \code{>}, and \code{>=}
}

Bitwise Operators
\itemize{
\item \code{bitwNot}, \code{bitwAnd}, \code{bitwOr} and \code{bitwXor}
\item \code{bitwShiftL} and \code{bitwShiftR}
}
}
\seealso{
\code{\link[base]{Arithmetic}}, \code{\link[base]{Comparison}},
\code{\link[base]{bitwNot}}.
}
