% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.r
\name{to_arcpy}
\alias{to_arcpy}
\alias{from_arcpy}
\title{R to/from ArcGIS Object}
\usage{
to_arcpy(x, ...)

from_arcpy(x, ...)
}
\arguments{
\item{x}{A raster or feature.}

\item{...}{Reservered for future expansion.}
}
\value{
For \code{to_arcpy()}, an ArcGIS raster or feature layer.
For \code{from_arcpy()}, a \code{SpatRaster} or \code{sf} object.
}
\description{
Convert rasters and features between R and ArcGIS.
}
