% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{use_arcpy}
\alias{use_arcpy}
\alias{require_arcpy}
\title{Connect to ArcGIS}
\usage{
use_arcpy(env_name, quietly = FALSE)

require_arcpy(env_name, quietly = TRUE)
}
\arguments{
\item{env_name}{The name of the ArcGIS Pro Conda environment,
or the path to the ArcGIS Desktop Python binary.}

\item{quietly}{If \code{TRUE}, suppress outputs.}
}
\description{
Connect to an ArcGIS Python environment.
}
\details{
The following settings from \code{\link[=options]{options()}} will be used
to populate missing arguments:
\itemize{
\item \code{arcpy.env}: (Character) The ArcGIS Pro Conda environment or path
to ArcGIS Desktop Python binary.
}
}
\examples{
\dontrun{
# Try to autodetect ArcGIS Pro installation
use_arcpy()

# connect to the 32-bit ArcGIS Desktop 10.2 Python environment
use_arcpy("C:/Python27/ArcGIS10.2")

}

}
