% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcpy.r
\docType{package}
\name{arcpy-package}
\alias{arcpy-package}
\alias{arcpy}
\title{Interface to ArcPy}
\format{
An object of class \code{python.builtin.module}.
}
\usage{
arcpy
}
\description{
An interface to the ArcGIS \code{arcpy} Python module via the
R-Python interface provided by \code{reticulate}. Loading the
packages exposes the \code{arcpy} python module for accessing the
ArcGIS geoprocessor. See the vignettes to get started.
}
\details{
\code{arcpy} uses the following \code{\link[=options]{options()}} to configure behavior:

\itemize{
\item \code{arcpy.env}: The name of an ArcGIS conda environment, or the
path to an ArcGIS Desktop Python binary.
}
}
\examples{
\dontrun{
use_arcpy()
arcpy$GetInstallInfo()
mygis = arcgis$gis$GIS()
}

}
\seealso{
\code{\link{use_arcpy}}, \code{\link{RasterAlgebra}}
}
\keyword{datasets}
