% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcpy.r
\docType{package}
\name{arcpy-package}
\alias{arcpy-package}
\alias{arcpy}
\title{Interface to ArcPy}
\description{
An interface to the ArcGIS \code{arcpy} Python module via the
R-Python interface provided by \code{reticulate}. Loading the
packages exposes the \code{arcpy} and \code{arcgis} python modules for
accessing the ArcGIS geoprocessor.
See the vignettes to get started.
}
\seealso{
\code{\link[=install_arcpy]{install_arcpy()}}
}
