% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcpy.r
\docType{package}
\name{arcpy-package}
\alias{arcpy-package}
\title{R Interface to ArcPy}
\description{
An R interface for the ArcGIS Python module \code{arcpy} via the
R-Python interface provided by reticulate. See the vignettes to get
started.
}
\seealso{
\code{\link{use_ArcGIS}}
}
