% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcpy.r
\docType{package}
\name{arcpy-package}
\alias{arcpy-package}
\alias{arcpy}
\title{Interface to ArcPy}
\format{An object of class \code{python.builtin.module}.}
\usage{
arcpy
}
\description{
An interface to the ArcGIS \code{arcpy} Python module via the
R-Python interface provided by \code{reticulate}. Loading the
packages exposes the \code{arcpy} python module for accessing the
ArcGIS geoprocessor. See the vignettes to get started.
}
\examples{
\dontrun{
use_ArcGIS()
arcpy$GetInstallInfo()
}

}
\seealso{
\code{\link{use_ArcGIS}}, \code{\link{RasterAlgebra}}
}
\keyword{datasets}
