% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.r
\name{da_drop}
\alias{da_drop}
\title{Table Row Removal with arcpy.da}
\usage{
da_drop(table.path, rows)
}
\arguments{
\item{table.path}{The file path to the table.}

\item{rows}{The row indexes to drop.}
}
\value{
(Invisible) The path to the table, i.e. \code{table.path}.
}
\description{
Drop records from a table (e.g. attribute table of a layer)
with the arcpy.da module.
}
\examples{
\dontrun{
layer = "path/to/table"
fields = c("VALUE", "NOTE")
d = da_read(layer, fields)
drop.rows = which(d$VALUE == 5)
da_drop(layer, drop.rows)
}

}
