% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.r
\name{da_insert}
\alias{da_insert}
\title{Table Insertion with arcpy.da}
\usage{
da_insert(table.path, d)
}
\arguments{
\item{table.path}{The file path to the table.}

\item{d}{The data to write to \code{table.path}, with the same number
of rows as the table. Column names must match field names
of the table.}
}
\value{
(Invisible) The path to the table, i.e. \code{table.path}.
}
\description{
Insert records into a table (e.g. attribute table of a layer)
with the arcpy.da module.
}
\examples{
\dontrun{
layer = "path/to/table"
fields = c("VALUE", "NOTE")
d = da_read(layer, fields)
add.d = data.frame(VALUE = 5, NOTE = "NEW",
  stringsAsFactors = FALSE)
da_insert(layer, add.d)
}

}
