% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcpy-methods.r
\name{RasterAlgebra}
\alias{RasterAlgebra}
\alias{+.python.builtin.Raster}
\alias{-.python.builtin.Raster}
\alias{*.python.builtin.Raster}
\alias{/.python.builtin.Raster}
\alias{^.python.builtin.Raster}
\alias{\%/\%.python.builtin.Raster}
\alias{\%\%.python.builtin.Raster}
\alias{<.python.builtin.Raster}
\alias{>.python.builtin.Raster}
\alias{<=.python.builtin.Raster}
\alias{>=.python.builtin.Raster}
\alias{!=.python.builtin.Raster}
\alias{==.python.builtin.Raster}
\alias{!.python.builtin.Raster}
\alias{&.python.builtin.Raster}
\alias{|.python.builtin.Raster}
\alias{xor.python.builtin.Raster}
\alias{bitwNot.python.builtin.Raster}
\alias{bitwAnd.python.builtin.Raster}
\alias{bitwOr.python.builtin.Raster}
\alias{bitwXor.python.builtin.Raster}
\alias{bitwShiftL.python.builtin.Raster}
\alias{bitwShiftR.python.builtin.Raster}
\title{Arcpy Raster Algebra}
\usage{
\method{+}{python.builtin.Raster}(a, b)

\method{-}{python.builtin.Raster}(a, b)

\method{*}{python.builtin.Raster}(a, b)

\method{/}{python.builtin.Raster}(a, b)

\method{^}{python.builtin.Raster}(a, b)

\method{\%/\%}{python.builtin.Raster}(a, b)

\method{\%\%}{python.builtin.Raster}(a, b)

\method{<}{python.builtin.Raster}(a, b)

\method{>}{python.builtin.Raster}(a, b)

\method{<=}{python.builtin.Raster}(a, b)

\method{>=}{python.builtin.Raster}(a, b)

\method{!=}{python.builtin.Raster}(a, b)

\method{==}{python.builtin.Raster}(a, b)

\method{!}{python.builtin.Raster}(a)

\method{&}{python.builtin.Raster}(a, b)

\method{|}{python.builtin.Raster}(a, b)

\method{xor}{python.builtin.Raster}(a, b)

\method{bitwNot}{python.builtin.Raster}(a)

\method{bitwAnd}{python.builtin.Raster}(a, b)

\method{bitwOr}{python.builtin.Raster}(a, b)

\method{bitwXor}{python.builtin.Raster}(a, b)

\method{bitwShiftL}{python.builtin.Raster}(a, n)

\method{bitwShiftR}{python.builtin.Raster}(a, n)
}
\arguments{
\item{a, b}{arcpy raster objects.}

\item{n}{non-negative integer.}
}
\description{
Operators for arcpy raster algebra.
}
\details{
The following operators are supported:

Arithmetic Operators:
\itemize{
\item \code{+}, \code{-}, and\code{/}
\item \verb{**} and \code{^}
\item \code{\%\%} and \code{\%/\%}
}

Logical Operators:
\itemize{
\item \code{!}, \code{&}, \code{|}, and \code{xor}
\item \code{!=} and \code{==}
\item \code{<}, \code{<=}, \code{>}, and \code{>=}
}

Bitwise Operators
\itemize{
\item \code{bitwNot}, \code{bitwAnd}, \code{bitwOr} and \code{bitwXor}
\item \code{bitwShiftL} and \code{bitwShiftR}
}
}
\examples{
\dontrun{
use_arcpy()
arcpy$env$workspace = tempdir()
arcpy$env$scratchWorkspace = tempdir()
arcpy$CheckOutExtension("Spatial")

cellSize = 2
outExtent = arcpy$Extent(0, 0, 250, 250)
raster1 = arcpy$sa$CreateConstantRaster(12, "FLOAT",
  cellSize, outExtent)

raster2 = arcpy$sa$CreateConstantRaster(6, "FLOAT",
  cellSize, outExtent)

raster1 + raster2
raster1 * raster2
raster1 ^ raster2
raster1 \%\% raster2

raster1 > raster2
(raster1 > raster2) & (raster2 > 7)
(raster1 > raster2) | (raster2 > 7)

arcpy$CheckInExtension("Spatial")
}

}
\seealso{
\code{\link[base]{Arithmetic}},
\code{\link[base]{Comparison}}, \code{\link[base]{bitwNot}}.
}
