% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{use_ArcGIS}
\alias{use_ArcGIS}
\title{Connect to ArcGIS Python}
\usage{
use_ArcGIS(pro = TRUE, condaenv, installpath, python)
}
\arguments{
\item{pro}{If \code{TRUE}, attempt to connect to an ArcGIS Pro
Python distribution. Otherwise, attempt to connect to an ArcGIS
Desktop Python distribution. Ignored if \code{python} is provided.}

\item{condaenv}{(ArcGIS Pro only). The ArcGIS Pro Conda environment
to use. If missing, the function will attempt to automatically
detect the default Conda environment.}

\item{installpath}{(ArcGIS Pro only). The ArcGIS Pro installation
directory. If missing, the function will attempt to
automatically detect the installation directory.}

\item{python}{(ArcGIS Desktop only) Path to the ArcGIS Desktop
Python binary. If missing, the function will attempt to
automatically detect an appropriate Python installation.}
}
\value{
No return value.
}
\description{
Connect to the ArcGIS Python environment.
}
\examples{
\dontrun{
# Try to autodetect ArcGIS Pro installation
use_ArcGIS()

# connect to the 32-bit ArcGIS Desktop 10.2 Python environment
use_ArcGIS("C:/Python27/ArcGIS10.2")

}

}
\seealso{
\code{\link{find_ArcGIS}}
}
