% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.r
\name{use_ArcGIS}
\alias{use_ArcGIS}
\title{Connect to ArcGIS Python}
\usage{
use_ArcGIS(pro, pythonpath)
}
\arguments{
\item{pro}{If \code{TRUE}, attempt to connect to an ArcGIS Pro
Python distribution. Otherwise, attempt to connect to an ArcGIS
Desktop Python distribution. Ignored if \code{python} is provided.}

\item{pythonpath}{(ArcGIS Desktop only) Path to the ArcGIS Desktop
Python binary. If missing, the function will attempt to
automatically detect an appropriate Python installation.}

\item{condaenv}{(ArcGIS Pro only). The ArcGIS Pro Conda environment
to use. If missing, the function will attempt to automatically
detect the default Conda environment.}

\item{installpath}{(ArcGIS Pro only). The ArcGIS Pro installation
directory. If missing, the function will attempt to
automatically detect the installation directory.}
}
\description{
Connect to the ArcGIS Python environment.
}
\details{
The following settings from \code{options()} will be used
to populate missing arguments:
\itemize{
\item \code{arcpy.pro}: (Logical) use ArcGIS Pro.
\item \code{arcpy.installpath}: (Character) The install path of ArcGIS Pro.
\item \code{arcpy.condaenv}: (Character) The path to the ArcGIS Pro Conda environment.
\item \code{arcpy.pythonpath}: (Character) The path to the ArcGIS Desktop Python installation.
}
}
\note{
(For ArcGIS Pro) The ArcGIS Pro bin folder (typically
\verb{C:\\Program Files\\ArcGIS\\Pro\\bin}) must be added to the PATH
variable before the Conda Environment is loaded. For users with
adminstrator privileges, \code{withr::with_path()} is used to temporarily
add the ArcGIS Pro bin folder to the path. Users without
adminstrator privileges may need to manually add the bin folder to
their user PATH variable before starting an R session.
}
\examples{
\dontrun{
# Try to autodetect ArcGIS Pro installation
use_ArcGIS()

# connect to the 32-bit ArcGIS Desktop 10.2 Python environment
use_ArcGIS("C:/Python27/ArcGIS10.2")

}

}
\seealso{
\code{\link{find_ArcGIS}}
}
